requirements.txt
coverage==5.0.3
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
dj-database-url==0.5.0
gunicorn==20.0.4
psycopg2==2.8.4
requests==2.22.0
whitenoise==5.0.1
